Add Virtual Steering Wheel Controls
Implement a virtual steering-wheel using the X and Z accelerometer axis on the 3DS.
The steering wheel value is sent as a separate axis (using the linux "wheel"
absolute-axis ID) in the gamepad's HID report.  Substituting the regular "analog X"
axis for this wheel axis from programs running on the server, they can then be
controlled by holding the 3DS as a steering wheel.